+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.0.1 (2026-05-16)
+
+- Fix: Add support for type-only packages, as indicated by only having a `types` key in `package.json`'s `exports` field
+- Fix: `.d.ts` files are no longer being incorrectly flagged as "this file has no compilation, so skip compile mapping step"
+
 ## 1.0.0 (2026-05-16)
 
 - Initial release